GUI: Fix type safety for 32/64-bit compatibility
This patch fixes critical type safety issues to ensure
correct behavior on both 32-bit and 64-bit systems.

Changes:
1. FTL interface: uint32_t -> uintptr_t for addresses
   - Fixes address truncation on 64-bit systems
   - Maintains compatibility on 32-bit embedded systems
   - Updated all FTL implementations (win32/keil)

2. Pointer-to-int conversions: int -> size_t
   - Prevents data loss in callback parameters
   - Uses semantically correct type for indices

3. Unaligned memory access: direct cast -> memcpy
   - Fixes potential crashes on ARM architectures
   - Safe on all platforms, compiler optimizes well

4. Macro fixes: unsigned long -> uintptr_t
   - Ensures cross-platform consistency
   - Removes Windows/Linux differences

All changes are binary compatible with 32-bit ARM
embedded systems. Compiled and tested successfully.

+++ example/application/screen_466_466/app_main.c
@@ -926,7 +926,7 @@ static void press_setting_cb_key(void *obj, gui_event_t e, void *param)
     GUI_UNUSED(e);
     GUI_UNUSED(param);
     // gui_obj_t *o = GUI_BASE(obj);
-    int index = (int)(uintptr_t)param;
+    size_t index = (size_t)param;
     gui_img_set_image_data((gui_img_t *)obj,
                            (void *)FILE_POINTER(key_highlight_array[index]));
     IMPORT_GUI_TOUCHPAD
@@ -1202,7 +1202,7 @@ static void release_setting_cb_key(void *obj, gui_event_t e, void *param)
     GUI_UNUSED(e);
     GUI_UNUSED(param);
     // gui_obj_t *o = GUI_BASE(obj);
-    int index = (int)(uintptr_t)param;
+    size_t index = (size_t)param;
     gui_img_set_image_data((gui_img_t *)obj,
                            (void *)FILE_POINTER(key_array[index]));
